+++ backend/tarsy/agents/iteration_controllers/react_stage_controller.py
@@ -12,7 +12,8 @@ from tarsy.models.unified_interactions import LLMMessage
 from .base_iteration_controller import IterationController, IterationContext
 
 if TYPE_CHECKING:
-    pass
+    from tarsy.integrations.llm.client import LLMClient
+    from tarsy.agents.prompt_builder import PromptBuilder
 
 logger = get_module_logger(__name__)
 
@@ -25,7 +26,8 @@ class ReactStageController(IterationController):
     analysis, providing incremental insights while accumulating data for next stages.
     """
     
-    def __init__(self, llm_client, prompt_builder):
+    def __init__(self, llm_client: 'LLMClient', prompt_builder: 'PromptBuilder'):
+        """Initialize with proper type annotations."""
         self.llm_client = llm_client  
         self.prompt_builder = prompt_builder
